requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
anyio==4.4.0
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
construct==2.10.68
construct-typing==0.5.6
discord.py==2.4.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonalias==0.1.1
multidict==6.0.5
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
solana==0.34.3
solders==0.21.0
typing_extensions==4.12.2
urllib3==2.2.2
websockets==11.0.3
yarl==1.9.4
